Add metadata utility to replace next/head across app router pages
