+++ src/app/(home)/team/page.tsx
@@ -1,0 +1,68 @@
+/* eslint-disable @next/next/no-img-element */
+import { departmentsData } from "components/team/team-data";
+import { DepartmentsEnum } from "components/team/team-interfaces";
+import TeamCard from "components/team/TeamCard";
+import TeamHero from "components/team/TeamHero";
+
+const TeamPage = () => {
+	return (
+		<div className="team">
+			<TeamHero heroTitle="Our Team" />
+
+			<main className="container-fluid team-categories">
+				<div className="strokes">
+					<img src="/images/Team/Ellipse-right.png" alt="" className="ellipse-right" />
+					<img src="/images/Team/Ellipse-left.png" alt="" className="ellipse-left" />
+				</div>
+
+				<div className="desc">
+					<p className="rochester-p">Leadership</p>
+					<h4>Pastors & Elders</h4>
+				</div>
+				<section className="department container-sm">
+					{departmentsData
+						.filter((elem) => elem.department == DepartmentsEnum.LEADERSHIP)
+						.map((elem) => (
+							<TeamCard team={elem} key={elem.id} />
+						))}
+				</section>
+
+				<div className="desc">
+					<p id="rochester-p">Staff</p>
+					<h4>Technicals</h4>
+				</div>
+				<section className="department container-sm">
+					{departmentsData
+						.filter((elem) => elem.department == DepartmentsEnum.TECHNICALS)
+						.map((elem) => (
+							<TeamCard team={elem} key={elem.id} />
+						))}
+				</section>
+
+				<div className="desc">
+					<h4>Music Department</h4>
+				</div>
+				<section className="department container-sm">
+					{departmentsData
+						.filter((elem) => elem.department == DepartmentsEnum.MUSIC_DEPARTMENT)
+						.map((elem) => (
+							<TeamCard team={elem} key={elem.id} />
+						))}
+				</section>
+
+				<div className="desc">
+					<h4>Workers</h4>
+				</div>
+				<section className="department container-sm">
+					{departmentsData
+						.filter((elem) => elem.department == DepartmentsEnum.WORKERS)
+						.map((elem) => (
+							<TeamCard team={elem} key={elem.id} />
+						))}
+				</section>
+			</main>
+		</div>
+	);
+};
+
+export default TeamPage;
